исправления

+++ common/src/main/kotlin/entities/NetPacket.kt
@@ -2,5 +2,5 @@ package online.fivem.common.entities
 
 class NetPacket(
 	val data: Any,
-	val key: Long? = null
+	val key: Double? = null
 )
