+++ common/src/main/kotlin/events/EstablishConnection.kt
@@ -4,5 +4,5 @@ package online.fivem.common.events
  * говорит клиенту, что он должен в дальнейшем использовать key для передачи пакетов
  */
 class EstablishConnection(
-	val key: Long
+	val key: Double
 )
